backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
numpy==2.0.1
opencv-python==4.10.0.84
scikit-learn==1.5.1
joblib==1.4.2
pillow==10.4.0

torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
